Fix logic to count facets in the response.
When a search returns zero results, the facets bar on the left should
not be displayed. The facets collection suppresses the
'search:facetsReturned' message if the count of the facets is 0.

The logic that obtained this count was counting the wrong thing; it
counted the number of different items under each facet heading, rather
than summing their counts. In other words, for the "Spatial Scope"
facet, it saw "Local", "Regional", and "Global" in the response, so the
count was 3, rather than summing how many results were associated with
each facet value.

[Trello #1042]
[Pivotal 77580716]