Fix home page link for executing NASA ECHO search.
In search-solr-tools, the long name for the NASA ECHO data center was
changed to have "(ECHO)" at the end. In the search-interface, values
defined in src/conf/acadis/appConfig-(integration|qa|staging).js define
the `expectedDataCenters`, for rendering the number of datasets we have
for each data center on the home page.

We need to have these values accessible outside of solr in the event
that we cannot access from one of those external data centers; we want
to be able to indicate that the ADE would normally have data from that
source, but none is available right now, therefore we cannot rely solely
on the information returned by solr to list each of the expected data centers.

- add "(ECHO)" to the end of the long name in the expectedDataCenters
  configuration across all environments

https://bitbucket.org/nsidc/search-solr-tools/commits/1bc9dd924c869cc217a47ef7b398987246e26c4d

[Pivotal 77676646]