Add browser test for datacenter links on ADE home.
On the ADE home page, we list all of the data centers we harvest data
from, and show how many datasets from each center we have. The text
specifying the number of datasets is also a link that executes a faceted
search, to take the user to those datasets.

Since the solr harvest can fail to ingest data sets from a given source
if it is down at the time of the harvest, we store a list of the
expected data centers in
src/conf/acadis/appConfig-(integration|qa|staging).js, which configures
the search that is executed from the ADE home page. If the data center
long or short names in solr change, the links on the ADE home page will
break.

- add browser tests to be certain the links on the ADE home page execute
  the searches and get the right number of results