Add test for always searching from URL params.
Test coverage for code added in 7b87aaa; ensure that a search event is
triggered when searching from the parameters in the URL, even if the
initial facets query returned zero results. This prevents the page from
displaying the loading results view and hanging forever--the user will
always be taken to results or the messgae that there are no results.

[Trello #1071]
[Pivotal 77888518]