Fix display of USGS results w/o temporal coverage.
USGS results were being rendered such that there was a "Temporal
Coverage:" label but no value, and the "Data Center:" label beneath it
was being indented to where the metdata values are supposed to be
rendered (screenshot attached to trello card on the Platypus team
board). The date information coming from the OpenSearch response for
these datasets is just an empty string for both start and end date.

This fix does not solve that metadata issue (which may need to be fixed
by adjusting the solr harvest or contacting USGS), just the rendering.

- filter ranges from the model so that only those with a valid start
  date (as determined by MomentJS) are rendered in the template

- refactoring in NsidcTemporalMetadataView.js:

    - move the logic for choosing which template will be rendered into
      `intialize()`

    - set the template to `this.template`, following the suggestion at
      http://backbonejs.org/#View-template

    - cleanup of the variable that handles the `dateRanges` from the
      model: call it `ranges` rather than `range`; don't check for it to
      be defined before cheking is length, since `_.filter` will return
      `[]` if `ranges` is undefined

- modify the test for when the start date is empty so that it expects
  nothing to be rendered

- fix some specs in NsidcTemporalMetadataView_spec.js that were using
  incorrect properties in the Backbone model

[Trello #1974]
[Pivotal 77643776]