Reset geobbox in `SearchParamsModel.resetCriteria`
All the other search criteria are reset correctly, but somehow the geo
bounding box got left out. This was causing a problem with the live data
center counts in the ADE where resetting a search returned the user to
the home page, but the search for the counts still had the spatial
restraint; see the [Pivotal story][0] for more detail.

[0]: https://www.pivotaltracker.com/story/show/89554538