Use different versions of OpenLayers for ADE/NSIDC Search
The OpenLayers upgrade is causing a strange issue for NSIDC Search where
the home page only has the search terms field; the spatial and temporal
fields are not rendered, nor is the "Data Sets for Researchers"
description text.

We wanted to upgrade OpenLayers for the ADE to fix the mouseover
coordinate issue (which is still present in NSIDC Search), so make the
different apps use different versions of OpenLayers.