Fixed NSIDC Search map with strange character in coords
The ACADIS fix had been implemented previously, but it needed to be
implemented for NSIDC Search as well.  Also added a "size" option to
the map initialization, as for some reason the new openlayers looks for
it instead of tileSize (not sure why this change wasn't needed for the
ACADIS fix).

Also made a slight update to the DEVELOPMENT documentation.